requirements.txt
absl-py==2.1.0
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blinker==1.8.2
boto3==1.34.127
botocore==1.34.127
Brotli==1.1.0
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.10.0
exceptiongroup==1.2.1
filelock==3.15.1
fire==0.6.0
Flask==3.0.3
Flask-SocketIO==5.3.6
fonttools==4.53.0
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.1
h11==0.14.0
huggingface-hub==0.23.4
hypothesis==6.103.2
idna==2.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
mutagen==1.47.0
netifaces==0.10.6
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.82
opencv-python-headless==4.8.0.74
packaging==24.1
pafy==0.5.5
pandas==2.2.2
pillow==10.3.0
pluggy==1.5.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pybboxes==0.1.6
pybind11==2.12.0
pycryptodomex==3.20.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-magic==0.4.27
python-socketio==5.11.2
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.32
rq==1.16.2
s3transfer==0.10.1
sahi==0.11.16
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
socketio==0.2.1
sortedcontainers==2.4.0
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tomli==2.0.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.32
ultralytics-thop==0.2.8
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
yolov5==7.0.13
youtube-dl==2021.12.17
yt-dlp==2024.5.27
